refactor : 퀴즈 카테고리 api 관리자용 분리
